Restructure the directories after adding new components
