+++ src/App.js
@@ -1,8 +1,8 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Mission from './components/view/missions/Mission';
 import Rocket from './components/view/rockets/Rocket';
-import Nav from './components/view/Nav';
-import Profile from './components/view/Profile';
+import Nav from './components/view/nav/Nav';
+import Profile from './components/view/profile/Profile';
 
 const App = () => (
   <div className="container">
