+++ src/components/view/nav/Nav.module.css
@@ -2,7 +2,6 @@
   display: flex;
   justify-content: space-between;
   align-items: baseline;
-  align-items: center;
   flex-wrap: wrap;
   padding: 20px 0;
   background: #fff;
@@ -51,7 +50,7 @@
 }
 
 .logo {
-  width: 70px;
-  height: 50px;
+  width: 75px;
+  height: 45px;
   margin-right: 1rem;
 }
